Refactor into more hooks
