+++ editor.planx.uk/src/@planx/components/Checklist/Public/helpers.ts
@@ -2,6 +2,7 @@ import { type Group } from "@planx/components/Checklist/model";
 import { Store } from "pages/FlowEditor/lib/store";
 
 import { Option } from "../../shared";
+import { FormikProps } from "formik";
 
 export function toggleInArray<T>(value: T, arr: Array<T>): Array<T> {
   return arr.includes(value)
